backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
stripe==7.4.0
openai==1.3.5
qrcode[pil]==7.4.2
aiofiles==23.2.1
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2
pydantic==2.5.0
pydantic[email]==2.5.0
redis==5.0.1
celery==5.3.4
boto3==1.33.5
pillow==10.1.0
httpx==0.25.2
cryptography==41.0.7
colorama==0.4.6
pydantic-settings==2.1.0
